testing around if a fixed/narrow prior on Mstar can help solve outshining
Idea was to provide a strong prior on Mstar when fitting PSB SEDs, which hopefully can pin down the amount of mass formed in the older stellar population, assuming the Mstar in the starburst is recovered with much higher accuracy. The strong Mstar prior can possibly be informed by dynamical mass estimations from measuring circular velocities or velocity dispersions of the galaxies from maps. This test does not seem to work out in the end...